Adjust plugin to use ts if sw and sts are 0.
Was irking the hell out of me that some ftplugins will set sw to 0, so
I would end up pressing tab at the end of a line and getting only
a single space. Rather than modify a whole lot of other plugins, we'll
just have this use tabstop when all else fails.